fix: make note-not-found reflect the global 404
Note-misses under /notes/* served a bare hand-rolled HTML page instead of
the branded global 404. Reuse the global-404 design for both locales while
keeping the case-redirect and fuzzy "did you mean…?" behavior.

- Parameterize NotFoundContent (heading/description/seoTitle/noindex props +
  a suggestion slot); add noindex support to Seo.
- Add a noteNotFound dictionary block (en/id) and prerendered not-found pages
  at /notes/not-found and /id/notes/not-found carrying suggestion-injection
  tokens; emit a localized /id/notes-index.json.
- Extract the edge logic to functions/_lib/notes-router.ts; make the en and
  new id /notes functions thin locale-configured wrappers that fetch the
  branded page, inject the suggestion, and return 404.
- Extend the middleware 404 guard to skip /id/notes/* so the id branch keeps
  its branded page instead of reverting to the global 404.
